// src/Services/Games/Lotofacil/Types/LotofacilType.ts
//Tipo do dado do sorteio da lotofacil
export interface LotofacilType {
    id?: string
    conc: number
    date: string
    ball1: number
    ball2: number
    ball3: number
    ball4: number
    ball5: number
    ball6: number
    ball7: number
    ball8: number
    ball9: number
    ball10: number
    ball11: number
    ball12: number
    ball13: number
    ball14: number
    ball15: number
}

//Tipo retornado do excelManager
export interface LotofacilExcel {
    Concurso: number,
    Data: string,
    'bola 1': number,
    'bola 2': number,
    'bola 3': number,
    'bola 4': number,
    'bola 5': number,
    'bola 6': number
    'bola 7': number
    'bola 8': number
    'bola 9': number
    'bola 10': number
    'bola 11': number
    'bola 12': number
    'bola 13': number
    'bola 14': number
    'bola 15': number
}